Added a tool for testing the HTMLification system. HTMLification is the process of converting raw text into HTML that is meaningful. Version 0.2 of docu is going to have much better HTMLification, and so this testing tool should aid the coding of this area.
The tool requires two files to be put into the 'htmlificaiton' directory for each test. The .txt file is what the user would enter into a DocBlock comment or the like, and the .htm file is the 
expected output after HTMLification.

This commit also includes two tests (both of which currently fail).

If end users would like to submit tests, that would be greatly appreciated.